+++ src/app/(home)/@chats/_components/chat-list.component.tsx
@@ -1,5 +1,5 @@
 "use client"
-import ChatPage from "./chat.component";
+import ChatListItem from "./chat-list-item.component";
 import { IUserChatRawDto } from "@/dtos/chats/chats.dtos";
 import { useAppDispatch, useAppSelector } from "@/store/hooks";
 import { fetchChats } from "@/store/thunks/chats.thunks";
@@ -34,7 +34,7 @@ export default function ChatList({ className }: { className?: string }) {
         <List disablePadding={true}>
             {
                 chats.map((chat, idx) =>
-                    <ChatPage key={idx} chat={chat} />)
+                    <ChatListItem key={idx} chat={chat} />)
             }
         </List>
     )
