+++ src/app/@modal/(.)chat/[chatId]/page.tsx
@@ -1,5 +1,5 @@
 
-import Chat from "@/app/chat/(chatid)/[chatId]/page";
+import Chat from "@/app/chat/[chatId]/page";
 import Modal from "@/app/_components/modal.component";
 
 export default function ChatModal(props: { params: { chatId: string } }) {
