+++ src/app/chat/[chatId]/_components/input-bar.component.tsx
