+++ src/app/chat/[chatId]/_components/message-bubbles/message-bubble-typing.component.tsx
