+++ src/app/chat/[chatId]/page.tsx
@@ -1,9 +1,9 @@
 import ChatSection from "./_components/chat-section.component";
-import * as chatServerService from "@/server-services/chats/chats.server-service";
+import * as chatServerService from "@/store/services/chat/chat.service";
 import { notFound } from "next/navigation";
 
 export default async function Chat(props: { params: { chatId: string } }) {
-    const chat = await chatServerService.getChat(props.params.chatId);    
+    //const chat = await chatServerService.getChat(props.params.chatId);    
 
     if (!chat.chatId) {
         return notFound();
